refactor(scheduler): rename m_utils and m_utils_time_zone (drop m_ prefix)
