+++ packages/devextreme/js/__internal/scheduler/r1/timezone_calculator/utils.ts
@@ -1,4 +1,4 @@
-import timeZoneUtils from '../../m_utils_time_zone';
+import timeZoneUtils from '../../utils_time_zone';
 import { TimeZoneCalculator } from './calculator';
 
 export const createTimeZoneCalculator = (
