+++ packages/devextreme/js/__internal/scheduler/workspaces/timeline.ts
@@ -17,7 +17,7 @@ import {
   GROUP_ROW_CLASS,
 } from '../classes';
 import tableCreatorModule, { type GroupRows } from '../m_table_creator';
-import timezoneUtils from '../m_utils_time_zone';
+import timezoneUtils from '../utils_time_zone';
 import HorizontalShader from '../shaders/current_time_shader_horizontal';
 import type { ResourceLoader } from '../utils/loader/resource_loader';
 import { getFirstVisibleDate } from '../utils/skipped_days';
