+++ packages/devextreme/js/__internal/scheduler/workspaces/view_model/m_date_header_data_generator.ts
@@ -1,7 +1,6 @@
 import dateUtils from '@js/core/utils/date';
 import type { DateHeaderData } from '@ts/scheduler/types';
 
-import timeZoneUtils from '../../m_utils_time_zone';
 import {
   formatWeekdayAndDay,
   getDisplayedCellCount,
@@ -11,6 +10,7 @@ import {
   isTimelineView,
 } from '../../r1/utils/index';
 import { VIEWS } from '../../utils/options/constants_view';
+import timeZoneUtils from '../../utils_time_zone';
 import type { ViewDataProviderExtendedOptions } from './m_types';
 
 export class DateHeaderDataGenerator {
