+++ packages/devextreme/js/__internal/scheduler/workspaces/view_model/m_grouped_data_map_provider.ts
@@ -2,7 +2,7 @@ import dateUtils from '@js/core/utils/date';
 import { dateUtilsTs } from '@ts/core/utils/date';
 import { isDateAndTimeView } from '@ts/scheduler/r1/utils/index';
 
-import timezoneUtils from '../../m_utils_time_zone';
+import timezoneUtils from '../../utils_time_zone';
 
 const toMs = dateUtils.dateToMilliseconds;
