+++ packages/devextreme/js/__internal/scheduler/workspaces/view_model/m_view_data_generator.ts
@@ -3,7 +3,6 @@ import { dateUtilsTs } from '@ts/core/utils/date';
 import type { GroupLeaf } from '@ts/scheduler/utils/resource_manager/types';
 
 import { HORIZONTAL_GROUP_ORIENTATION } from '../../constants';
-import timezoneUtils from '../../m_utils_time_zone';
 import {
   calculateCellIndex,
   calculateDayDuration,
@@ -22,6 +21,7 @@ import {
   getVisibleDaysOfWeek,
   isDateSkipped,
 } from '../../utils/skipped_days';
+import timezoneUtils from '../../utils_time_zone';
 import type {
   ViewCellDataSimple,
   ViewCellGeneratedData,
