+++ packages/devextreme/js/__internal/scheduler/workspaces/work_space_indicator.ts
@@ -10,7 +10,7 @@ import type { OptionChanged } from '@ts/core/widget/types';
 import { getToday } from '@ts/scheduler/r1/utils/index';
 
 import { HEADER_CURRENT_TIME_CELL_CLASS } from '../classes';
-import timezoneUtils from '../m_utils_time_zone';
+import timezoneUtils from '../utils_time_zone';
 import SchedulerWorkSpace, {
   type WorkspaceOptionsInternal,
 } from './m_work_space';
