+++ packages/devextreme/js/__internal/scheduler/workspaces/work_space_month.ts
@@ -6,7 +6,7 @@ import { hasWindow } from '@js/core/utils/window';
 import { DateTableMonthComponent } from '@ts/scheduler/r1/components/index';
 import { formatWeekday, monthUtils } from '@ts/scheduler/r1/utils/index';
 
-import { utils } from '../m_utils';
+import { utils } from '../utils';
 import { VIEWS } from '../utils/options/constants_view';
 import type { ViewDateGenerationOptions } from './m_work_space';
 import SchedulerWorkSpace from './work_space_indicator';
