+++ packages/devextreme/testing/tests/DevExpress.ui.widgets.scheduler/integration.appointmentsVertical.tests.js
@@ -7,7 +7,7 @@ import translator from 'common/core/animation/translator';
 import fx from 'common/core/animation/fx';
 import { DataSource } from 'common/data/data_source/data_source';
 import dataUtils from 'core/element_data';
-import timeZoneUtils from '__internal/scheduler/m_utils_time_zone';
+import timeZoneUtils from '__internal/scheduler/utils_time_zone';
 
 import 'fluent_blue_light.css!';
 import '__internal/scheduler/m_scheduler';
